ENTM 6820 Project
In this repository, you can find two folders. One folder includes all class assignments for ENTM 6820; the other includes the files for the reproducible workflow assignment.

The data for this project is 8 weeks worth of crape myrtle bark scale population counts from CMBS infested crape myrtles in the greenhouse. Adult and Crawler counts were taken weekly over 2 months on 36 trees with 6 different treatments. All data are physical, in-situ counts, taken by hand. CMBS adults and crawlers were counted on 10 cm of 3 random branches. The most common control method is drench applications of the neonicotinoid of imidacloprid. The goal of the experiment was to determine if limonene, a botanical, had similar or better control of CMBS than common chemical insecticides including pyriproxyfen, dinotefuran, and imidacloprid, as well as a water treatment and a non-treated control. Based on previous unpublished data, we hypothesized that limonene would meet or exceed our expectations for the control of CMBS.

In this repository you can find the data collected as well as an R Markdown file containing code, analysis, and figures for presentation.